+++ tests/rx/rx_test.hpp
@@ -15,7 +15,7 @@ protected:
   void RunFor(std::chrono::duration<Rep, Period> duration,
               std::function<void()> f = nullptr) {
     auto const then{std::chrono::system_clock::now() + duration};
-    while (std::chrono::system_clock::now() < then) f ? f() : _mock.execute();
+    while (std::chrono::system_clock::now() < then) f ? f() : _mock.receive();
   }
 
   NiceMock<RxMock> _mock;